changes for 64-bit machines
SCCS-vsn: sys/sys/resource.h 8.3
SCCS-vsn: sys/sys/tablet.h 8.4
SCCS-vsn: sys/sys/time.h 8.2

+++ usr/src/sys/sys/resource.h
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)resource.h	8.2 (Berkeley) %G%
+ *	@(#)resource.h	8.3 (Berkeley) %G%
  */
 
 #ifndef _SYS_RESOURCE_H_
@@ -66,8 +66,8 @@ struct	rusage {
 #define	RLIM_INFINITY	(((u_quad_t)1 << 63) - 1)
 
 struct orlimit {
-	long	rlim_cur;		/* current (soft) limit */
-	long	rlim_max;		/* maximum value for rlim_cur */
+	int32_t	rlim_cur;		/* current (soft) limit */
+	int32_t	rlim_max;		/* maximum value for rlim_cur */
 };
 
 struct rlimit {
@@ -77,8 +77,8 @@ struct rlimit {
 
 /* Load average structure. */
 struct loadavg {
-	fixpt_t ldavg[3];
-	long fscale;
+	fixpt_t	ldavg[3];
+	int32_t fscale;
 };
 
 #ifdef KERNEL
